#33 refactor : ChallengeDetail 공용화, navigation 수정
