+++ app/src/main/java/com/aoztg/greengrim/presentation/ui/NavigationExtensions.kt
@@ -4,7 +4,12 @@ import androidx.navigation.NavController
 import com.aoztg.greengrim.MainNavDirections
 
 
-internal fun NavController.toAttendCheck(){
+internal fun NavController.toAttendCheck() {
     val action = MainNavDirections.actionGlobalToAttendCheckFragment()
+    this.navigate(action)
+}
+
+internal fun NavController.toChallengeDetail(id: String) {
+    val action = MainNavDirections.actionGlobalToChallengeDetailFragment(id)
     this.navigate(action)
 }
