+++ app/src/main/java/com/aoztg/greengrim/presentation/ui/global/challengedetail/ChallengeDetailViewModel.kt
@@ -1,4 +1,4 @@
-package com.aoztg.greengrim.presentation.ui.challenge.detail
+package com.aoztg.greengrim.presentation.ui.global.challengedetail
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
